Add scaffold for next step.
Which will be make is so that when we click on add to cart button on
a sponsorship we get redirected to the sponsor node add page or if
a sponsorship product is already in our cart we get redirect to the
sponsor node edit page.

Need to take care of OK now, so I'll complete later.